feat(wallet): verifySignature
